Revise QWinJumpList::addXxx()
Return the created item so that the user can adjust its properties
of desired. Also, limit the amount of addLink() overloads so that
the user has a chance to memorize the arguments. Offer only the
most common arguments (executable & arguments) and the rest less
common attributes may be then set via the returned pointer.

Change-Id: Ic1575ac441fd44b5a7915bbae473970effeeb023
Reviewed-by: Laszlo Papp <lpapp@kde.org>
Reviewed-by: Ivan Vizir <define-true-false@yandex.com>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@digia.com>