Refactor QQuickIconLoader.
Instead using a QObject with a single icon member variable
that needs to be instantiated for each task bar button, introduce a static
load function that takes a member function pointer accepting a QVariant.
For synchronous requests, invoke it immediately; connect it to a slot
for network requests.

Task-number: QTBUG-36730
Change-Id: Ic70422d98d0bc3d486b32fb3534400b9fff8d324
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>