Doc: Fix documentation warnings related to Clang-enabled QDoc
Since Qt 5.11, QDoc uses Clang to parse C++ documentation. On platforms
where Qt Windows Extras does not build, we still need to be able to
build the documentation.

To enable this, create a custom module header and include paths for
Clang-QDoc, used only in documentation builds, and add dummy typedefs
for Windows-specific types for Clang to see.

Change-Id: I641570c80b6cc948c92c404d8d90ad1a3045925c
Reviewed-by: Martin Smith <martin.smith@qt.io>