Update commercial license header
Updated header.COMM to the files in tqtc-qtwinextras. Examples,
documentation files or tests are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4166
Change-Id: I735c35b3d1d0782de411fb9d2368d22b3853b21b
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>